+++ metalearners/tlearner.py
@@ -1,12 +1,21 @@
 # Copyright (c) QuantCo 2024-2024
 # SPDX-License-Identifier: BSD-3-Clause
 
+from collections.abc import Mapping, Sequence
+
 import numpy as np
 from joblib import Parallel, delayed
 from typing_extensions import Self
 
-from metalearners._typing import Matrix, OosMethod, Scoring, Vector
-from metalearners._utils import index_matrix
+from metalearners._typing import Matrix, OosMethod, Scoring, Vector, _ScikitModel
+from metalearners._utils import (
+    check_spox_installed,
+    copydoc,
+    index_matrix,
+    infer_input_dict,
+    infer_probabilities_output,
+    warning_experimental_feature,
+)
 from metalearners.cross_fit_estimator import OVERALL
 from metalearners.metalearner import (
     NUISANCE,
@@ -17,6 +26,7 @@ from metalearners.metalearner import (
     _fit_cross_fit_estimator_joblib,
     _ModelSpecifications,
     _ParallelJoblibSpecification,
+    get_overall_estimators,
 )
 
 
@@ -140,3 +150,53 @@ class TLearner(_ConditionalAverageOutcomeMetaLearner):
             is_treatment_model=False,
             feature_set=self.feature_set[VARIANT_OUTCOME_MODEL],
         )
+
+    def _necessary_onnx_models(self) -> dict[str, list[_ScikitModel]]:
+        return {
+            VARIANT_OUTCOME_MODEL: get_overall_estimators(
+                self._nuisance_models[VARIANT_OUTCOME_MODEL]
+            )
+        }
+
+    @copydoc(MetaLearner._build_onnx, sep="")
+    def _build_onnx(self, models: Mapping[str, Sequence], output_name: str = "tau"):
+        """In the TLearner case, the necessary models are:
+
+        * ``"variant_outcome_model"``
+        """
+        warning_experimental_feature("_build_onnx")
+        check_spox_installed()
+        import spox.opset.ai.onnx.v21 as op
+        from onnx.checker import check_model
+        from spox import build, inline
+
+        self._validate_feature_set_none()
+        self._validate_onnx_models(models, set(self._necessary_onnx_models().keys()))
+
+        input_dict = infer_input_dict(models[VARIANT_OUTCOME_MODEL][0])
+
+        if not self.is_classification:
+            model_output_name = models[VARIANT_OUTCOME_MODEL][0].graph.output[0].name
+        else:
+            _, model_output_name = infer_probabilities_output(
+                models[VARIANT_OUTCOME_MODEL][0]
+            )
+
+        output_0 = inline(models[VARIANT_OUTCOME_MODEL][0])(**input_dict)[
+            model_output_name
+        ]
+
+        variant_cates = []
+        for m in models[VARIANT_OUTCOME_MODEL][1:]:
+            variant_output = inline(m)(**input_dict)[model_output_name]
+            variant_cate = op.sub(variant_output, output_0)
+            variant_cates.append(
+                op.unsqueeze(
+                    variant_cate,
+                    axes=op.constant(value_int=1),
+                )
+            )
+        cate = op.concat(variant_cates, axis=1)
+        final_model = build(input_dict, {output_name: cate})
+        check_model(final_model, full_check=True)
+        return final_model