+++ CHANGELOG.rst
@@ -16,6 +16,11 @@ Changelog
 
 * Add :class:`metalearners.utils.FixedBinaryPropensity`.
 
+* Added ``_build_onnx`` to :class:`metalearners.MetaLearner` abstract class and implement it
+  for :class:`metalearners.TLearner`, :class:`metalearners.XLearner`, :class:`metalearners.RLearner`
+  and :class:`metalearners.DRLearner`.
+
+* Added ``_necessary_onnx_models`` to :class:`metalearners.MetaLearner`.
 
 0.8.0 (2024-07-22)
 ------------------